Optimization vs Accuracy
Decreasing the rate at which calculations are done every few frames, decreasing image sizes and computation rates, and the option to toggle visualization on and off for when we wont be monitoring outputs such as during the Gallery showing, all in the multicam_scaffold.py

As for homography_test.py, I wanted to match the confidence and image size rates to that in the scaffold so they stay aligned during runtime. Half is there to spare my laptop and that of others who attempt to test this when the scale gets much larger.