Reorganizing repo and creating systemd services
Reorganized the file structure of the repo so it is much cleaner and better for expanding on.
Also created an install.sh script. This script will automatically install all python packages that are needed, as well as moving the scripts from the location it was cloned, to the expected directory at /opt/p2bp/camera.
Finally, the first test systemd service was created for the camera_scanner.py. This service should start when the Jetson boots and it will ensure that the tracking scripts can access each camera.